CCI update
UItooltipbox was fixed to not have a standard vertical orientation for
the main body
The Server Browser was updated to reflect the change in UItooltipbox,
and additionally, a variable has been made local.

CCI has been updated to make use of UItooltipbox as well, and now has
basic functionality that actually gives it a reason for existing. It's
still immensely barebones but better than to not work at all.