another boring night
lib.cfg : each created UI now assigns a timestamp of when it opened to a
variable with the UI's name

helpers.cfg : more work done on linear gradients using the coswave
function. UIcoswave was added to simplify Y value acquisition from
within UIs, and there's now coswave.INT/RGB24 which do the same but for
colors - they can also work outside of an UI though. Prettified the
"coswave" UI to better showcase this in action, and ultimately no user
variables would be required besides the wavelength.