I don't remember
Lots of cleanups on the bind system, renamed slightly and introduced descriptors that are to be used to describe what a bind does - the rest is up to the coder.
Changed vars for default map lists due to some weird bug that caused them to not get saved for some people (???)
Added get(auth/dead/lagged) - how did i miss those
UIentcolors, UIentcell and UIhudcells moved to main lib
The styling got a little more stylish by adding the ability to pulse an outline upon hover/hold by the 5th param of default style suites, can be disabled with -1. Customized references of these styles across the menus have been updated to match this addition where appropriate. 
Turns out "grassstep" works the opposite way and so the name was changed accordingly to "Blade Sparsity".
First edits to the "keybinds" menu (formerly "controls") to get the descriptors implemented. I'm lazy and I work slow.