Oops
Fixed the selection var for the geo_prefab menu, it was using the one from the mapmodels menu before.

Also, since it's now part of the HUD interface (and technically the scoreboard is too but that will change at a later time), it was moved to the hud folder. Also made sure to hide it when the user exits editmode automatically.

The gcd and lcm functions were moved to the common.cfg, more appropriate to have them there. Considering getting rid of the whole helpers.cfg file.