moar updates
stdlib.cfg - Changed .crosshairsize default to 0 as it was conflicting with the first use of the followplayer function. Working on improving that overall.

default.cfg - Moved the new msc_delta function here and out from stdedit.cfg. It's been expanded to encompass more behaviors rather than leave the code clutter in the binds themselves.

style.cfg - Repositioned the condition to account for some edge cases.

helpers.cfg - Sanity check for the RGB variant as well.

common.cfg - Slight improvements to the followplayer function. Now accounts for an empty playerlist to turn itself off, and has a dedicated function for changing the followed player. Still in the middle of making it better but it works at least.

map_browser.cfg - Working on an entirely new version of it, in a similar style to the texture and model browser. Does not work fully yet. 

texture/mapmodel_browser.cfg & geo_prefab.cfg - All of them use a UIbuttonarea now to toggle the hover sound. They also account for map_browser menu too, and use the hidden .x & .y vars to simplify some code in generic size maths.