I have no good title.
Some all rounder updates.

One important addition, getselface and getselentface commands. This way we know which side of a selection or entity is the primary. Can be used to write more scripts that depends on orientation, such as vcommands. Will get on to those next.

New Recorder UI has been added for movies. Accessible from the main menu, currently still a WIP and prone to further improvements down the line.

main.cfg: Cleaning up a few stuffs and adding said shortcut to the new menu.
options.cfg: Adding in some UI-specific switches, removed the Sound on/off switch (figure people won't really touch it) and added additional settings for audio.
lib.cfg: UIbasicfield now has a dedicated child for when the var is null, addition of UIvsliderslim just so we have that.

The rest are some minor edits for continuity.