rudimentary vdirdelta
First iteration of the "vdirdelta" command which is meant to transform directional voffset/vscroll input to the appropriate cardinal direction when the top/bottom face of a cube is selected so that they all move to the actual direction you're looking.

Unfortunately it turns out that different sides of a cube result in an inverted X or Y direction for some goddamn reason so it needs to be remade to account for those.

default.cfg -- Edited to include the `vdirdelta` command.
stdedit.cfg -- Foreshadowing of full 3D rotation for applicable entities with the R key and selected face.
stdlib.cfg -- Remove the sanitization code for keybinds. It's causing too many problems and nearly nothing gets saved right.

octaedit.cpp -- cleaned up `getselface` and `getselentface` so that they're simpler and more flexible.