Slider adjustments (mainly)
-- stdlib.cfg 
Addition of `!iskeyheld` for convenience.

-- lib.cfg
Triangles cannot be clamped to size so that part was commented out.  Cleaned up UIhbarslider to give it a more consistent feel that makes using it a bit less clunky, also cleared up its code a bit for consistency. The UIoptionslider itself was also refactored, though whether the efficiency of the code was increased is unknown. The code is cleaner now, and it no longer requires the list to be separated in two parts.

-- helpers.cfg
Addition of a small Snake game demo, will relocate it somewhere else down the line.

-- common.cfg
Addition of pad0, pad0f, and precf commands for string/number formatting.

-- server_browser.cfg
The new font is a smidge wider than before, so I had to rename  button to no longer stretch the frame and enlarge the whole UI.

-- options.cfg
Adjusted all option sliders to the new format.