It's been a long time coming.
Though not entirely done yet with the new featured addition, I reckoned it was a good time to save up the accumulated changes. Here we go! 

stdlib.cfg -- Addition of the keyboard hook variables, and modification of keyswitch to call upon a new function that will hijack keyboard input.

stdedit.cfg -- Removal of some useless vars, `entdirection` (the function called with R+scroll in edit)  can now rotate certain entities in all their possible directions based on the selected side of the entity's bounding box. Experimental `suncamangle`,  `suncamdist` and `sundir` UI, that's a WIP and on hold until I can learn to math.

defaults.cfg -- Fixed `vdirdelta` parameters on arrow keys, re-added the missing binds for BACKSPACE (not sure when/why those disappeared).

helpers.cfg -- `coswave` and `sinwave` were missing float operators, added those. Other coswave functions were modified to ensure the X value is always positive (countermeasure for millis overflowing to negative). The biggest change here is the *finally* fixed `vdirdelta` function. After a ton of trial and error, blood and tears over how "annoyingly" a certain face has flipped coordinates on only one axis, it now works flawlessly and one can at last offset or scroll a texture in the direction they want! Constricted by a fixed directional input no more! Also renamed a couple experimental UIs, and you'll also notice a function meant for translating. I will return to that in the future.

style.cfg -- Just some comments added.

lib.cfg -- This place is a bit of a mess at present.. the biggest additions being `UIadvancedfieldLA` and `UIadvancedfieldRA`, for left and right aligned text respectively. These new fields are practically fully configurable and look so much better than the `UIbasicfield` that was sparingly used thus far due to its scaling issues, lackluster interaction control and styling constraints. There's also  `.UI_KBreset` to init the new field's values, `.UI_KBinput` to handle the keyboard hijack behavior (configurable for more field types!!), `UIcaret` which is what its name implies and I need to make it local to the fields, it's useless outside of them, `UI_KBclose` which is a failsafe to kill a field's jack should the UI that it resided in goes away (still working on that).. Other changes include an OOB var for each UI (Out Of Bounds -- basically a glorified uihover check, might remove this after since it was an experiment) as well as a "shelf" var. That is used for situations where you might want to make some UI element *slide in*  ;-)  to view when you hover over a box. Used in the character profile UI, but uiclip restrictions kind of hold us back now as it won't clip in the -X and -Y directions BLEH. Last minor changes are the functions for the shelf behavior I just mentioned, the display value of sliders being *above* the slider line, the removal of `diagline` from the tooltipbox.. 

main.cfg -- Returned the button for player profile configuration and renamed it. Removed the entire segment on the right side of the menu that previously held an alt version of player-setup.cfg's contents. Added a call to define the `KBhook` UI at the end and start it up immediately. It is invisible and it only exists to allow left/right clicks outside of an actual UI to close an active field.

player_setup.cfg -- Bit of an overhaul. The weapon buttons now slide in from the bottom of the model preview. The name was supposed to slide in from the top, but due to the aforementioned issue with uiclip, it will instead always display right over the model preview. That part is a bit of a mess due to experiments with the fields, will be cleaned up in a followup commit. The loop for the color boxes was cleaned up just a bit. 

recorder.cfg -- Do not reset moview/movieh when the UI opens if a recording is still going on. 

server_browser.cfg -- Removed diagline call for an inline instead to avoid errors.

That's about it for now! More changes will follow as we go.